Create test scripts and byte-width test inputs
A builddict test script still needs to be written, likewise
inputs for other symbol widths and suffix array widths.
Testing builddict may be a bit tricky because of the pseudo-random
generator; how strictly are the C++ standard library PRNGs
specified? Perhaps just testing that the output is of an expected
size is enough. I also need to write tests for random-position
decompression, which is still rlzunparse's least tested feature.